refactor(types): split cockpit evidence DTOs
