+++ docs/architecture-consolidation-plan.md
@@ -69,6 +69,7 @@ fixtures:
 | Format fun outputs | `crates/tokmd-format/src/fun/mod.rs` and `crates/tokmd-format/src/fun/` | `mod.rs` 18; OBJ owner 136; MIDI owner 199 | OBJ code-city rendering and MIDI rendering now live in format-specific owner modules while the root fun module preserves public re-exports and feature-gated behavior |
 | Format redaction utilities | `crates/tokmd-format/src/redact/mod.rs` and `crates/tokmd-format/src/redact/` | `mod.rs` 251; safe extension owner 44 | Redaction hashing and path redaction remain in the public root module while safe file-extension preservation policy lives in its own private owner module; keep this under the `format_redaction_scan_args` proof scope |
 | Core context/handoff DTO contracts | `crates/tokmd-types/src/lib.rs`, `crates/tokmd-types/src/inventory.rs`, `crates/tokmd-types/src/context.rs`, and `crates/tokmd-types/src/diff.rs` | `lib.rs` 453; inventory owner 453; context owner 432; diff owner 85 | Lang/module/export/run inventory receipt DTOs, context receipts, context bundle manifests, handoff manifests, token estimation/audit DTOs, artifact/capability DTOs, and diff receipt DTOs now live in dedicated owners while root re-exports preserve the public contract |
+| Cockpit receipt DTO contracts | `crates/tokmd-types/src/cockpit.rs` and `crates/tokmd-types/src/cockpit/` | receipt/metrics/review/trend coordinator 352; evidence gate owner 241 | Cockpit receipt, metric, review-plan, and trend DTOs stay re-exported from the `cockpit` module while evidence gate DTOs live in `cockpit/evidence.rs`. Preserve the `tokmd_cockpit` proof scope and cockpit schema behavior. |
 | Envelope contract | `crates/tokmd-envelope/src/lib.rs`, `crates/tokmd-envelope/src/artifact.rs`, `crates/tokmd-envelope/src/capability.rs`, `crates/tokmd-envelope/src/finding.rs`, `crates/tokmd-envelope/src/gate.rs`, `crates/tokmd-envelope/src/ffi.rs`, and `crates/tokmd-envelope/src/findings.rs` | `lib.rs` 395; artifact owner/tests 85; capability owner/tests 96; finding DTO owner/tests 289; gate owner/tests 154; FFI envelope helper 299; finding ID registry 84 | Keep `lib.rs` as the root sensor report contract and public re-export surface; keep artifact DTO builders and tests in `artifact.rs`, capability availability DTO builders and tests in `capability.rs`, portable finding DTO builders and tests in `finding.rs`, gate result/item DTO builders and tests in `gate.rs`, FFI response extraction in `ffi.rs`, and finding ID constants in `findings.rs`. Preserve the `tokmd_envelope_contract` proof scope. |
 | Settings contracts | `crates/tokmd-settings/src/lib.rs`, `crates/tokmd-settings/src/commands.rs`, `crates/tokmd-settings/src/config.rs`, `crates/tokmd-settings/src/profile.rs`, and `crates/tokmd-settings/src/scan.rs` | `lib.rs` 36; command settings owner 249; TOML config owner 259; profile owner 32; scan owner 61 | Keep `lib.rs` as root re-export glue; keep workflow command settings in `commands.rs`, TOML configuration contracts in `config.rs`, legacy profile compatibility in `profile.rs`, and shared scan settings in `scan.rs`. Preserve the `tokmd_settings_contract` proof scope. |
 | Git adapter boundary | `crates/tokmd-git/src/lib.rs`, `crates/tokmd-git/src/command.rs`, `crates/tokmd-git/src/intent.rs`, and `crates/tokmd-git/src/refs.rs` | git adapter coordinator 251; command/env isolation owner 82; intent classifier owner 165; ref-resolution owner 242 | Keep public history, diff, base-ref, and intent APIs stable from `lib.rs`; keep subprocess environment scrubbing and `git_cmd` construction in `command.rs`; keep Conventional Commit and keyword-based intent classification in `intent.rs`; keep revision existence, base-ref fallback resolution, and env-ref sanitization in `refs.rs`. Preserve `git_subprocess_boundary`, `tokmd_git_intent_classification`, and `tokmd_git_ref_resolution` as git adapter files move. |
